docs(mkdocs): clean content and add short demo (#45)
This pull request updates the documentation for Tombo to improve clarity
and consistency in the feature descriptions. The main changes are the
removal of emoji icons from feature headings and lists, and the addition
of a more detailed demo section on the index page. These updates make
the documentation more professional and easier to read.